cmd/trace: add goroutine label events in UI, filter by labels
Add goroutine label instantaneous events to traces in the UI. Also, add
query parameters to highlight goroutines based on their labels. The "lk"
and "lv" parameters specify label keys and values, respectively. An
empty value parameter means any value is acceptable. The parameters can
be repeated, and they will be matched up pairwise in the order they
appear. For example, this query says the key "keyA" should have value
"X", and that "keyB" should be present with any value:

    "https://127.0.0.1:6060/trace?lk=keyA&lv=X&lk=keyB&lv="

Events tied to goroutines which had the given labels with the
given values will be higlighted, and all other events will be greyed
out.

TODO: link to this somehow from the main tracer page.

Change-Id: I02e7f876cfd17c5f1ff4429b6e1455c1be7b65d9